+++ bootstrap/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.5.3
 
+- Fixed bootstrap logging filter
 - Update `avail-light-core` to 1.2.9
 
 ## [0.5.2](https://github.com/availproject/avail-light/releases/tag/avail-light-bootstrap-v0.5.2) - 2025-04-09